[coop] Move syscall abort mechanism to dedicated files
