+++ mono/utils/mono-threads.c
@@ -589,6 +589,7 @@ mono_threads_init (MonoThreadInfoCallbacks *callbacks, size_t info_size)
 	mono_lls_init (&thread_list, NULL);
 	mono_thread_smr_init ();
 	mono_threads_init_platform ();
+	mono_threads_init_abort_syscall ();
 
 #if defined(__MACH__)
 	mono_mach_init (thread_info_key);
