+++ msvc/libmonoutils.vcxproj
@@ -63,8 +63,12 @@
     <ClCompile Include="..\mono\utils\mono-stdlib.c" />
     <ClCompile Include="..\mono\utils\mono-threads-mach.c" />
     <ClCompile Include="..\mono\utils\mono-threads-posix.c" />
+    <ClCompile Include="..\mono\utils\mono-threads-posix-signals.c" />
     <ClCompile Include="..\mono\utils\mono-threads-windows.c" />
     <ClCompile Include="..\mono\utils\mono-threads.c" />
+    <ClCompile Include="..\mono\utils\mono-threads-posix-abort-syscall.c" />
+    <ClCompile Include="..\mono\utils\mono-threads-mach-abort-syscall.c" />
+    <ClCompile Include="..\mono\utils\mono-threads-windows-abort-syscall.c" />
     <ClCompile Include="..\mono\utils\mono-time.c" />
     <ClCompile Include="..\mono\utils\mono-tls.c" />
     <ClCompile Include="..\mono\utils\mono-uri.c" />
@@ -129,6 +133,7 @@
     <ClInclude Include="..\mono\utils\mono-stdlib.h" />
     <ClInclude Include="..\mono\utils\mono-string.h" />
     <ClInclude Include="..\mono\utils\mono-threads.h" />
+    <ClCompile Include="..\mono\utils\mono-threads-posix-signals.h" />
     <ClInclude Include="..\mono\utils\mono-time.h" />
     <ClInclude Include="..\mono\utils\mono-tls.h" />
     <ClInclude Include="..\mono\utils\mono-uri.h" />
